requirements.txt
fastapi
uvicorn
youtube-transcript-api
langchain
langchain-community
langchain-huggingface
langchain-mistralai
langchain-text-splitters
faiss-cpu
tiktoken
sentence-transformers
python-dotenv
streamlit
requests